fix some admin stuff and much stuff in store

+++ bookingsystem/routes/bookings.js
@@ -3,24 +3,25 @@
 // model
 let Booking = require('../models/booking');
 let Cabin = require('../models/cabin');
+let auth = require('./auth');
 
 // POST
 module.exports.post = async (req, res) => {
-    console.log(req.body);
+
     try {
         let cabin = await Cabin.findById(req.body.cabin);
-            let eBooking = [];
+            let booking = [];
 
             for(i = 0; i < req.body.amount; i++) {
                 let booking = {
                     cabin: cabin,
                     code: uid(5),
                 }
-                eBooking.push(booking);
+                booking.push(booking);
             }
 
             // write bookings to Mongo
-            let resp = await Booking.create(eBooking);
+            let resp = await Booking.create(booking);
 
             // Send to FrontEnd
             res.status(200).send(resp);
@@ -29,13 +30,32 @@ module.exports.post = async (req, res) => {
         res.status(404).send(err.stack);
     }
 }
+module.exports.get = async (req, res) => {
 
+    try {
+
+        res.status(200).send( await Booking.find({}) );        
+    
+    } catch(err){
+    
+        res.status(500).send(err.stack);
+    
+    }
+
+}
 
 /* När vi behöver avboka en bokning */
 module.exports.delete = async (req, res) => {
     
     try {
-        res.status(200).send( await Cabin.findOneAndDelete({_id: req.params.bookingId}));
+
+        // Kolla om användaren är admin
+        if(auth.verifyToken(req.headers.authorization)){
+
+            res.status(200).send( await Cabin.findOneAndDelete({_id: req.params.bookingId}));
+
+        }
+    
     } catch(err) {
         res.status(404).send(err.stack);
     }
